мелкая оптимизация
