+++ src/mock/event-item.js
@@ -98,9 +98,9 @@ const EVENT_TYPES = {
 
 const descriptionProto = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus`;
 
-const DESCRIPTIONS = descriptionProto.split(`. `);
+const descriptions = descriptionProto.split(`. `);
 
-const getRandomDescription = () => getRandomInt(0, 5) ? getRandomArray(DESCRIPTIONS, 5).join(`. `) + `.` : ``;
+const getRandomDescription = () => getRandomInt(0, 5) ? getRandomArray(descriptions, 5).join(`. `) + `.` : ``;
 
 const getRandomPhotos = () => {
   const photos = [];
