+++ src/util.js
@@ -1,3 +1,5 @@
+import dayjs from "dayjs";
+
 export const getRandomInt = (min, max) => {
   return Math.floor(Math.random() * (max - min)) + min;
 };
@@ -14,4 +16,16 @@ export const getRandomArray = (array, count) => {
     randomArray[i] = getRandomEl(array);
   }
   return randomArray;
+};
+
+export const durationString = (beginning, ending) => {
+  const duration = dayjs(ending).diff(dayjs(beginning), `minute`);
+  const durationH = Math.floor(duration / 60);
+  const durationM = duration % 60;
+  let result = ``;
+  if (durationH) {
+    result += `${durationH}H `;
+  }
+  result += `${durationM}M`;
+  return result;
 };
